refactor: restructure md copy component naming
