+++ components/theme/notebook.tsx
@@ -7,7 +7,7 @@ import { Fragment, type HTMLAttributes, type ReactNode } from "react";
 
 import { cn } from "@/lib/theme/cn";
 
-import { MarkdownActionsClient } from "../docs/MarkdownActions/client";
+import { MarkdownActionsClient } from "../md-copy/client";
 import { Banner } from "./banner";
 import {
   SidebarLinkItem,
